Fixed a bug which prevented User.delete() from working. Intercom._call was not expecting the DELETE method when constructing request data parameters. Hence, the params were being skipped over and the request didn't correctly register with Intercom's API.
Added in intergration test to check this is working as indetented.
Before,
The user returned would be the 'bob@example.com' not the email specificed ('to-delete@example.com' in the tests case).
Now,
The user info returned corresponds to the email specified ('to-delete@example.com') as we would expect, as Intercom API 'Returns the deleted user on success' (as according to their docs).