Add a resource for the contacts endpoint (now called leads, but still contacts in the API)
Contacts are extremely similar to users. The main difference is that the user_id cannot be specified, it must be auto-assigned by Intercom. There are also a few properties that Users have but Contacts do not.

This creates the new resource, adds some tests for it, and does a minor refactor in the test utilities to make it easy to generate a contact payload without duplicating a whole bunch of code. I also noticed that the created_at field was specified twice in get_user(), so I removed the first instance (which was being ignored).