Add support for permanent user deletion.
What Intercom previously called `delete` for users was actually an
archive. They have now added a real deletion mechanism, and renamed
`delete` to `archive`.

For backwards compatibility, `delete` will continue to function the
same, performing an archive of the user. There is now a new `archive`
function (which simply calls `delete`), and a new `permanent_delete`,
which will request permanent deletion of the user.